Fix server error
After capturing the exception that was being thrown by the search route
when deploying to Heroku, it turned out Flask's jsonify method had a
limit on list length, if provided as an argument. Fixed it by returning
results from the API as a single object containing the list of results.